requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements-dev.txt

# Include base requirements
-r requirements.txt

# Code formatting
black==24.3.0
isort==5.13.2

# Linting
flake8==7.0.0
flake8-bugbear==24.2.6
flake8-comprehensions==3.14.0

# Type checking
mypy==1.9.0
django-stubs==4.2.7
djangorestframework-stubs==3.14.5

# Pre-commit hooks
pre-commit==3.7.0

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
coverage==7.4.4
pytest-mock==3.14.0
factory-boy==3.3.0
faker==25.2.0

# Security checking
safety==3.0.1
bandit==1.7.8
